Extract common functionality to the base class
Easier than updating the Fallback project host separately :)
