+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/ProjectSystem/DefaultWindowsRazorProjectHost.cs
@@ -10,7 +10,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Threading.Tasks.Dataflow;
 using Microsoft.AspNetCore.Razor.Language;
 using Microsoft.CodeAnalysis.Razor.Workspaces;
 using Microsoft.VisualStudio.LanguageServices;
@@ -31,8 +30,6 @@ internal class DefaultWindowsRazorProjectHost : WindowsRazorProjectHostBase
 {
     private const string RootNamespaceProperty = "RootNamespace";
     private readonly LanguageServerFeatureOptions _languageServerFeatureOptions;
-    private readonly List<IDisposable> _disposables = new();
-    private readonly Dictionary<ProjectConfigurationSlice, IDisposable> _projectSubscriptions = new();
 
     [ImportingConstructor]
     public DefaultWindowsRazorProjectHost(
@@ -59,113 +56,21 @@ internal class DefaultWindowsRazorProjectHost : WindowsRazorProjectHostBase
     {
     }
 
-    protected override async Task InitializeCoreAsync(CancellationToken cancellationToken)
+    protected override string[] GetRuleNames()
     {
-        await base.InitializeCoreAsync(cancellationToken).ConfigureAwait(false);
-
-        // CPS represents the various target frameworks that a project has in configuration groups, which are called "slices". Each
-        // slice represents a variation of a project configuration. So for example, a given multi-targeted project would have:
-        //
-        // Configuration    | Platform     | Configuration Groups (slices)
-        // -----------------------------------------------------
-        // Debug            | Any CPU      | net6.0, net7.0
-        // Release          | Any CPU      | net6.0, net7.0
-        //
-        // This subscription hooks to the ActiveConfigurationGroupSubscriptionService which will feed us data whenever a
-        // "slice" is added or removed, for the current active Configuration/Platform combination. This is a nice mix between
-        // not having too many things loaded (ie, we don't get 4 projects representing the full matrix of configurations) but
-        // still having distinct projects per target framework. If the user changes configuration from Debug to Release, we will
-        // get updates for the slices indicating that change, but within a specific configuration, both target frameworks will
-        // get updates for project changes, which means our data won't be stale if the user changes the active context.
-        //
-        // CPS also manages the slices themselves, and we get updates as they change. eg the first event we get has one target
-        // framework, then we get an update containing both. If the user adds one, we get another event.  Either way
-        // the event also gives us a datasource we can subscribe to in order to receive updates about that project. It is
-        // important that we maintain our list of subscriptions because if a slice is removed, we are responsible for cleaning
-        // up our resources.
-        //
-        // It's worth noting the events also give us a key, which is a list of "dimensions". We only care about the target framework
-        // but they could be strictly anything, and could change at any time. If they do, we'll get new events, so the easiest
-        // thing to do is just treat the key as an opaque object. CPS implements IEquatable on it, expressly for this purpose.
-        // We should not have any logic that depends on the contents of the key.
-        _disposables.Add(CommonServices.ActiveConfigurationGroupSubscriptionService.SourceBlock.LinkTo(
-            DataflowBlockSlim.CreateActionBlock<IProjectVersionedValue<ConfigurationSubscriptionSources>>(SlicesChanged, nameFormat: "Slice {0}"),
-            new DataflowLinkOptions() { PropagateCompletion = true }));
-
-        // Join, in the JTF sense, the ActiveConfigurationGroupSubscriptionService, to help avoid hangs in our OnProjectChangedAsync method
-        _disposables.Add(ProjectDataSources.JoinUpstreamDataSources(CommonServices.ThreadingService.JoinableTaskFactory, CommonServices.FaultHandlerService, CommonServices.ActiveConfigurationGroupSubscriptionService));
-    }
-
-    private void SlicesChanged(IProjectVersionedValue<ConfigurationSubscriptionSources> value)
-    {
-        // Create a new dictionary representing the subscriptions we know about at the start of the update. Data flow ensures
-        // this method will not be called in parallel.
-        var current = new Dictionary<ProjectConfigurationSlice, IDisposable>(_projectSubscriptions);
-
-        foreach (var (slice, source) in value.Value)
+        return new string[]
         {
-            if (!_projectSubscriptions.TryGetValue(slice, out var dataSource))
-            {
-                Assumes.False(current.ContainsKey(slice));
-
-                // This is a new slice that we didn't previously know about, either because its a new target framework, or how dimensions
-                // are calculated has changed. We simply subscribe to updates for it, and let our action block code handle whether the
-                // distinction is important. To put it another way, we may end up having multiple subscriptions and events that would be
-                // affect about the same project.razor.json file, but our event handling code ensures we don't handle them more than
-                // necessary.
-                var subscription = source.JointRuleSource.SourceBlock.LinkTo(
-                    DataflowBlockSlim.CreateActionBlock<IProjectVersionedValue<IProjectSubscriptionUpdate>>(OnProjectChangedAsync, "OnProjectChanged {0}"),
-                    initialDataAsNew: true,
-                    suppressVersionOnlyUpdates: true,
-                    ruleNames: new string[]
-                    {
-                        Rules.RazorGeneral.SchemaName,
-                        Rules.RazorConfiguration.SchemaName,
-                        Rules.RazorExtension.SchemaName,
-                        Rules.RazorComponentWithTargetPath.SchemaName,
-                        Rules.RazorGenerateWithTargetPath.SchemaName,
-                        ConfigurationGeneralSchemaName,
-                    });
-
-                _projectSubscriptions.Add(slice, subscription);
-            }
-            else
-            {
-                // We already know about this slice, so remove it from our "current" list, as we have nothing to do for it
-                Assumes.True(current.Remove(slice));
-            }
-        }
-
-        // Anything left in the current list must have been removed, so we dispose it
-        foreach (var (slice, subscription) in current)
-        {
-            Assumes.True(_projectSubscriptions.Remove(slice));
-
-            subscription.Dispose();
-        }
-    }
-
-    protected override async Task DisposeCoreAsync(bool initialized)
-    {
-        await base.DisposeCoreAsync(initialized).ConfigureAwait(false);
-
-        if (initialized)
-        {
-            foreach (var (slice, subscription) in _projectSubscriptions)
-            {
-                Assumes.True(_projectSubscriptions.Remove(slice));
-                subscription.Dispose();
-            }
-
-            foreach (var disposable in _disposables)
-            {
-                disposable.Dispose();
-            }
-        }
+            Rules.RazorGeneral.SchemaName,
+            Rules.RazorConfiguration.SchemaName,
+            Rules.RazorExtension.SchemaName,
+            Rules.RazorComponentWithTargetPath.SchemaName,
+            Rules.RazorGenerateWithTargetPath.SchemaName,
+            ConfigurationGeneralSchemaName,
+        };
     }
 
     // Internal for testing
-    internal async Task OnProjectChangedAsync(IProjectVersionedValue<IProjectSubscriptionUpdate> update)
+    internal override async Task OnProjectChangedAsync(IProjectVersionedValue<IProjectSubscriptionUpdate> update)
     {
         if (IsDisposing || IsDisposed)
         {
