+++ src/Razor/src/Microsoft.VisualStudio.LanguageServices.Razor/ProjectSystem/WindowsRazorProjectHostBase.cs
@@ -11,6 +11,7 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Threading.Tasks.Dataflow;
 using Microsoft.VisualStudio.LanguageServices;
 using Microsoft.VisualStudio.ProjectSystem;
 using Microsoft.VisualStudio.ProjectSystem.Properties;
@@ -29,7 +30,9 @@ internal abstract class WindowsRazorProjectHostBase : OnceInitializedOnceDispose
 
     private ProjectSnapshotManagerBase? _projectManager;
     private readonly Dictionary<string, HostDocument> _currentDocuments;
+    private readonly Dictionary<ProjectConfigurationSlice, IDisposable> _projectSubscriptions = new();
     protected readonly ProjectConfigurationFilePathStore ProjectConfigurationFilePathStore;
+    protected readonly List<IDisposable> Disposables = new();
 
     internal const string BaseIntermediateOutputPathPropertyName = "BaseIntermediateOutputPath";
     internal const string IntermediateOutputPathPropertyName = "IntermediateOutputPath";
@@ -104,12 +107,90 @@ internal abstract class WindowsRazorProjectHostBase : OnceInitializedOnceDispose
         return InitializeAsync();
     }
 
-    protected override Task InitializeCoreAsync(CancellationToken cancellationToken)
+    protected sealed override Task InitializeCoreAsync(CancellationToken cancellationToken)
     {
         CommonServices.UnconfiguredProject.ProjectRenaming += UnconfiguredProject_ProjectRenamingAsync;
 
+        // CPS represents the various target frameworks that a project has in configuration groups, which are called "slices". Each
+        // slice represents a variation of a project configuration. So for example, a given multi-targeted project would have:
+        //
+        // Configuration    | Platform     | Configuration Groups (slices)
+        // -----------------------------------------------------
+        // Debug            | Any CPU      | net6.0, net7.0
+        // Release          | Any CPU      | net6.0, net7.0
+        //
+        // This subscription hooks to the ActiveConfigurationGroupSubscriptionService which will feed us data whenever a
+        // "slice" is added or removed, for the current active Configuration/Platform combination. This is a nice mix between
+        // not having too many things loaded (ie, we don't get 4 projects representing the full matrix of configurations) but
+        // still having distinct projects per target framework. If the user changes configuration from Debug to Release, we will
+        // get updates for the slices indicating that change, but within a specific configuration, both target frameworks will
+        // get updates for project changes, which means our data won't be stale if the user changes the active context.
+        //
+        // CPS also manages the slices themselves, and we get updates as they change. eg the first event we get has one target
+        // framework, then we get an update containing both. If the user adds one, we get another event.  Either way
+        // the event also gives us a datasource we can subscribe to in order to receive updates about that project. It is
+        // important that we maintain our list of subscriptions because if a slice is removed, we are responsible for cleaning
+        // up our resources.
+        //
+        // It's worth noting the events also give us a key, which is a list of "dimensions". We only care about the target framework
+        // but they could be strictly anything, and could change at any time. If they do, we'll get new events, so the easiest
+        // thing to do is just treat the key as an opaque object. CPS implements IEquatable on it, expressly for this purpose.
+        // We should not have any logic that depends on the contents of the key.
+        Disposables.Add(CommonServices.ActiveConfigurationGroupSubscriptionService.SourceBlock.LinkTo(
+            DataflowBlockSlim.CreateActionBlock<IProjectVersionedValue<ConfigurationSubscriptionSources>>(SlicesChanged, nameFormat: "Slice {0}"),
+            new DataflowLinkOptions() { PropagateCompletion = true }));
+
+        // Join, in the JTF sense, the ActiveConfigurationGroupSubscriptionService, to help avoid hangs in our OnProjectChangedAsync method
+        Disposables.Add(ProjectDataSources.JoinUpstreamDataSources(CommonServices.ThreadingService.JoinableTaskFactory, CommonServices.FaultHandlerService, CommonServices.ActiveConfigurationGroupSubscriptionService));
+
         return Task.CompletedTask;
     }
+
+    private void SlicesChanged(IProjectVersionedValue<ConfigurationSubscriptionSources> value)
+    {
+        // Create a new dictionary representing the subscriptions we know about at the start of the update. Data flow ensures
+        // this method will not be called in parallel.
+        var current = new Dictionary<ProjectConfigurationSlice, IDisposable>(_projectSubscriptions);
+
+        foreach (var (slice, source) in value.Value)
+        {
+            if (!_projectSubscriptions.TryGetValue(slice, out var dataSource))
+            {
+                Assumes.False(current.ContainsKey(slice));
+
+                // This is a new slice that we didn't previously know about, either because its a new target framework, or how dimensions
+                // are calculated has changed. We simply subscribe to updates for it, and let our action block code handle whether the
+                // distinction is important. To put it another way, we may end up having multiple subscriptions and events that would be
+                // affect about the same project.razor.json file, but our event handling code ensures we don't handle them more than
+                // necessary.
+                var subscription = source.JointRuleSource.SourceBlock.LinkTo(
+                    DataflowBlockSlim.CreateActionBlock<IProjectVersionedValue<IProjectSubscriptionUpdate>>(OnProjectChangedAsync, "OnProjectChanged {0}"),
+                    initialDataAsNew: true,
+                    suppressVersionOnlyUpdates: true,
+                    ruleNames: GetRuleNames(),
+                    linkOptions: new DataflowLinkOptions() { PropagateCompletion = true });
+
+                _projectSubscriptions.Add(slice, subscription);
+            }
+            else
+            {
+                // We already know about this slice, so remove it from our "current" list, as we have nothing to do for it
+                Assumes.True(current.Remove(slice));
+            }
+        }
+
+        // Anything left in the current list must have been removed, so we dispose it
+        foreach (var (slice, subscription) in current)
+        {
+            Assumes.True(_projectSubscriptions.Remove(slice));
+
+            subscription.Dispose();
+        }
+    }
+
+    protected abstract string[] GetRuleNames();
+
+    internal abstract Task OnProjectChangedAsync(IProjectVersionedValue<IProjectSubscriptionUpdate> update);
 
     protected override async Task DisposeCoreAsync(bool initialized)
     {
@@ -124,6 +205,16 @@ internal abstract class WindowsRazorProjectHostBase : OnceInitializedOnceDispose
                     await UpdateAsync(UninitializeProjectUnsafe, CancellationToken.None).ConfigureAwait(false);
                 }
             }).ConfigureAwait(false);
+
+            foreach (var (slice, subscription) in _projectSubscriptions)
+            {
+                subscription.Dispose();
+            }
+
+            foreach (var disposable in Disposables)
+            {
+                disposable.Dispose();
+            }
         }
     }
